Added consumer idea and combine resources function in Resource

+++ correccion_y_locks/idle_game/src/utils/mod.rs
@@ -3,13 +3,13 @@ use rand::Rng;
 
 
 pub fn coin_flip() -> usize {
-    random_number_in(0, 1)
+    random_number_in(0, 1) as usize
 }
 
 pub fn extract_gold() -> usize {
-    random_number_in(MIN_GOLD, MAX_GOLD)
+    random_number_in(MIN_GOLD as isize, MAX_GOLD as isize) as usize
 }
 
-pub fn random_number_in(start: usize, finish: usize) -> usize {
+pub fn random_number_in(start: isize, finish: isize) -> isize {
     rand::thread_rng().gen_range(start..=finish)
 }
